[Refactor] 패키지 구조 변경에 따른 코드 리팩토링
