--- src/main/java/show/schedulemanagement/service/SignUpService.java
@@ -1,10 +1,0 @@
-package show.schedulemanagement.service;
-
-import show.schedulemanagement.domain.member.Member;
-import show.schedulemanagement.web.request.signup.SignUpRequestDto;
-
-public interface SignUpService {
-    Member signUp(SignUpRequestDto signUpRequestDto);
-    Boolean isDuplicatedEmail(String email);
-    Boolean isDuplicatedNickname(String nickname);
-}
